# boltdb shipper index gateway client improvements