--- a/pkg/storage/stores/shipper/gateway_client.go
+++ b/pkg/storage/stores/shipper/gateway_client.go
@@ -7,6 +7,7 @@ import (
 	"io"
 
 	"github.com/go-kit/log/level"
+	"github.com/grafana/dskit/concurrency"
 	"github.com/grafana/dskit/grpcclient"
 	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
@@ -21,7 +22,10 @@ import (
 	util_math "github.com/grafana/loki/pkg/util/math"
 )
 
-const maxQueriesPerGoroutine = 100
+const (
+	maxQueriesPerGrpc      = 200
+	maxConcurrentGrpcCalls = 10
+)
 
 type IndexGatewayClientConfig struct {
 	Address          string            `yaml:"server_address,omitempty"`
@@ -78,24 +82,17 @@ func (s *GatewayClient) Stop() {
 }
 
 func (s *GatewayClient) QueryPages(ctx context.Context, queries []chunk.IndexQuery, callback chunk.QueryPagesCallback) error {
-	errs := make(chan error)
-
-	for i := 0; i < len(queries); i += maxQueriesPerGoroutine {
-		q := queries[i:util_math.Min(i+maxQueriesPerGoroutine, len(queries))]
-		go func(queries []chunk.IndexQuery) {
-			errs <- s.doQueries(ctx, queries, callback)
-		}(q)
+	if len(queries) <= maxQueriesPerGrpc {
+		return s.doQueries(ctx, queries, callback)
 	}
 
-	var lastErr error
-	for i := 0; i < len(queries); i += maxQueriesPerGoroutine {
-		err := <-errs
-		if err != nil {
-			lastErr = err
-		}
+	jobsCount := len(queries) / maxQueriesPerGrpc
+	if len(queries)%maxQueriesPerGrpc != 0 {
+		jobsCount++
 	}
-
-	return lastErr
+	return concurrency.ForEachJob(ctx, jobsCount, maxConcurrentGrpcCalls, func(ctx context.Context, idx int) error {
+		return s.doQueries(ctx, queries[idx*maxQueriesPerGrpc:util_math.Min((idx+1)*maxQueriesPerGrpc, len(queries))], callback)
+	})
 }
 
 func (s *GatewayClient) doQueries(ctx context.Context, queries []chunk.IndexQuery, callback chunk.QueryPagesCallback) error {