@sandeepsukhani 2022-02-23
I just realised that the mutex change to protect the concurrent writes to grpc streams and this PR with increasing the number of calls per grpc call will go in the same release. I think there should be a difference of a release since the queriers are fast in rolling out, which means they will start sending 200 queries per grpc and would cause index gateways to concurrently process the queries which would cause a deadlock. I have changed the number back for both to 100 for now. We will change this again after a couple of releases and the performance for now is not that bad:

```
benchmark                    old allocs     new allocs     delta
Benchmark_MultiQueries-8     7798108907     3273959945     -58.02%

benchmark                    old allocs     new allocs     delta
Benchmark_MultiQueries-8     47906005       47648335       -0.54%

benchmark                    old bytes      new bytes      delta
Benchmark_MultiQueries-8     2694492664     2687846304     -0.25%
```